cloud-function/requirements.txt
google-api-core==2.0.1
google-cloud-pubsub==2.8.0